+++ lua/nvim-tree.lua
@@ -236,6 +236,20 @@ local function setup_autocommands(opts)
       end,
     })
   end
+
+  -- Handles event dispatch when tree is closed by `:q`
+  create_nvim_tree_autocmd("WinClosed", {
+    pattern = "*",
+    ---@param ev vim.api.keyset.create_autocmd.callback_args
+    callback = function(ev)
+      if not vim.api.nvim_buf_is_valid(ev.buf) then
+        return
+      end
+      if vim.api.nvim_get_option_value("filetype", { buf = ev.buf }) == "NvimTree" then
+        require("nvim-tree.events")._dispatch_on_tree_close()
+      end
+    end,
+  })
 end
 
 local DEFAULT_OPTS = { -- BEGIN_DEFAULT_OPTS
@@ -259,6 +273,7 @@ local DEFAULT_OPTS = { -- BEGIN_DEFAULT_OPTS
   view = {
     centralize_selection = false,
     cursorline = true,
+    cursorlineopt = "both",
     debounce_delay = 15,
     side = "left",
     preserve_window_proportions = false,
@@ -323,7 +338,10 @@ local DEFAULT_OPTS = { -- BEGIN_DEFAULT_OPTS
       hidden_placement = "after",
       diagnostics_placement = "signcolumn",
       bookmarks_placement = "signcolumn",
-      padding = " ",
+      padding = {
+        icon = " ",
+        folder_arrow = " ",
+      },
       symlink_arrow = " ➛ ",
       show = {
         file = true,
@@ -402,6 +420,7 @@ local DEFAULT_OPTS = { -- BEGIN_DEFAULT_OPTS
       warning = "",
       error = "",
     },
+    diagnostic_opts = false,
   },
   modified = {
     enable = false,
@@ -538,6 +557,7 @@ local ACCEPTED_TYPES = {
       "table",
       min = { "string", "function", "number" },
       max = { "string", "function", "number" },
+      lines_excluded = { "table" },
       padding = { "function", "number" },
     },
   },
@@ -597,6 +617,14 @@ local ACCEPTED_STRINGS = {
   },
 }
 
+local ACCEPTED_ENUMS = {
+  view = {
+    width = {
+      lines_excluded = { "root", },
+    },
+  },
+}
+
 ---@param conf table|nil
 local function validate_options(conf)
   local msg
@@ -605,15 +633,19 @@ local function validate_options(conf)
   ---@param def any
   ---@param strs table
   ---@param types table
+  ---@param enums table
   ---@param prefix string
-  local function validate(user, def, strs, types, prefix)
+  local function validate(user, def, strs, types, enums, prefix)
     -- if user's option is not a table there is nothing to do
     if type(user) ~= "table" then
       return
     end
 
-    -- only compare tables with contents that are not integer indexed
-    if type(def) ~= "table" or not next(def) or type(next(def)) == "number" then
+    -- we have hit a leaf enum to validate against - it's an integer indexed table
+    local enum_value = type(enums) == "table" and next(enums) and type(next(enums)) == "number"
+
+    -- only compare tables with contents that are not integer indexed nor enums
+    if not enum_value and (type(def) ~= "table" or not next(def) or type(next(def)) == "number") then
       -- unless the field can be a table (and is not a table in default config)
       if vim.tbl_contains(types, "table") then
         -- use a dummy default to allow all checks
@@ -627,27 +659,34 @@ local function validate_options(conf)
       if not FIELD_SKIP_VALIDATE[k] then
         local invalid
 
-        if def[k] == nil and types[k] == nil then
-          -- option does not exist
-          invalid = string.format("Unknown option: %s%s", prefix, k)
-        elseif type(v) ~= type(def[k]) then
-          local expected
+        if enum_value then
+          if not vim.tbl_contains(enums, v) then
+            invalid = string.format("Invalid value for field %s%s: Expected one of enum '%s', got '%s'", prefix, k,
+              table.concat(enums, "'|'"), tostring(v))
+          end
+        else
+          if def[k] == nil and types[k] == nil then
+            -- option does not exist
+            invalid = string.format("Unknown option: %s%s", prefix, k)
+          elseif type(v) ~= type(def[k]) then
+            local expected
 
-          if types[k] and #types[k] > 0 then
-            if not vim.tbl_contains(types[k], type(v)) then
-              expected = table.concat(types[k], "|")
+            if types[k] and #types[k] > 0 then
+              if not vim.tbl_contains(types[k], type(v)) then
+                expected = table.concat(types[k], "|")
+              end
+            else
+              expected = type(def[k])
             end
-          else
-            expected = type(def[k])
-          end
 
-          if expected then
-            -- option is of the wrong type
-            invalid = string.format("Invalid option: %s%s. Expected %s, got %s", prefix, k, expected, type(v))
+            if expected then
+              -- option is of the wrong type
+              invalid = string.format("Invalid option: %s%s. Expected %s, got %s", prefix, k, expected, type(v))
+            end
+          elseif type(v) == "string" and strs[k] and not vim.tbl_contains(strs[k], v) then
+            -- option has type `string` but value is not accepted
+            invalid = string.format("Invalid value for field %s%s: '%s'", prefix, k, v)
           end
-        elseif type(v) == "string" and strs[k] and not vim.tbl_contains(strs[k], v) then
-          -- option has type `string` but value is not accepted
-          invalid = string.format("Invalid value for field %s%s: '%s'", prefix, k, v)
         end
 
         if invalid then
@@ -657,14 +696,14 @@ local function validate_options(conf)
             msg = invalid
           end
           user[k] = nil
-        else
-          validate(v, def[k], strs[k] or {}, types[k] or {}, prefix .. k .. ".")
+        elseif not enum_value then
+          validate(v, def[k], strs[k] or {}, types[k] or {}, enums[k] or {}, prefix .. k .. ".")
         end
       end
     end
   end
 
-  validate(conf, DEFAULT_OPTS, ACCEPTED_STRINGS, ACCEPTED_TYPES, "")
+  validate(conf, DEFAULT_OPTS, ACCEPTED_STRINGS, ACCEPTED_TYPES, ACCEPTED_ENUMS, "")
 
   if msg then
     notify.warn(msg .. "\n\nsee :help nvim-tree-opts for available configuration options")
@@ -744,10 +783,7 @@ function M.setup(conf)
 
   setup_autocommands(opts)
 
-  if vim.g.NvimTreeSetup ~= 1 then
-    -- first call to setup
-    require("nvim-tree.commands").setup()
-  else
+  if vim.g.NvimTreeSetup == 1 then
     -- subsequent calls to setup
     M.purge_all_state()
   end