+++ lua/nvim-tree/actions/node/open-file.lua
@@ -2,6 +2,7 @@
 local lib = require("nvim-tree.lib")
 local notify = require("nvim-tree.notify")
 local utils = require("nvim-tree.utils")
+local full_name = require("nvim-tree.renderer.components.full-name")
 local view = require("nvim-tree.view")
 
 local M = {}
@@ -39,19 +40,13 @@ local function usable_win_ids()
     end
 
     local win_config = vim.api.nvim_win_get_config(id)
-    return id ~= tree_winid and win_config.focusable and not win_config.external or false
+    return id ~= tree_winid
+      and id ~= full_name.popup_win
+      and win_config.focusable
+      and not win_config.hide
+      and not win_config.external
+      or false
   end, win_ids)
-end
-
----Find the first window in the tab that is not NvimTree.
----@return integer -1 if none available
-local function first_win_id()
-  local selectable = usable_win_ids()
-  if #selectable > 0 then
-    return selectable[1]
-  else
-    return -1
-  end
 end
 
 ---Get user to pick a window in the tab that is not NvimTree.
@@ -203,7 +198,15 @@ local function open_file_in_tab(filename)
   if M.relative_path then
     filename = utils.path_relative(filename, vim.fn.getcwd())
   end
-  vim.cmd("tabe " .. vim.fn.fnameescape(filename))
+  vim.cmd.tabnew()
+  vim.bo.bufhidden = "wipe"
+  -- Following vim.fn.tabnew the # buffer may be set to the tree buffer. There is no way to clear the # buffer via vim.fn.setreg as it requires a valid buffer. Clear # by setting it to a new temporary scratch buffer.
+  if utils.is_nvim_tree_buf(vim.fn.bufnr("#")) then
+    local tmpbuf = vim.api.nvim_create_buf(false, true)
+    vim.fn.setreg("#", tmpbuf)
+    vim.api.nvim_buf_delete(tmpbuf, { force = true })
+  end
+  vim.cmd.edit(vim.fn.fnameescape(filename))
 end
 
 local function drop(filename)
@@ -246,9 +249,14 @@ local function get_target_winid(mode)
   local target_winid
   if not M.window_picker.enable or string.find(mode, "no_picker") then
     target_winid = lib.target_winid
-    -- first available window
-    if not vim.tbl_contains(vim.api.nvim_tabpage_list_wins(0), target_winid) then
-      target_winid = first_win_id()
+    local usable_wins = usable_win_ids()
+    -- first available usable window
+    if not vim.tbl_contains(usable_wins, target_winid) then
+      if #usable_wins > 0 then
+        target_winid = usable_wins[1]
+      else
+        target_winid = -1
+      end
     end
   else
     -- pick a window