+++ lua/nvim-tree/explorer/live-filter.lua
@@ -187,9 +187,11 @@ local function create_overlay(self)
   })
 
   if vim.fn.has("nvim-0.10") == 1 then
-    vim.api.nvim_set_option_value("modifiable", true, { buf = overlay_bufnr })
+    vim.api.nvim_set_option_value("modifiable", true,             { buf = overlay_bufnr })
+    vim.api.nvim_set_option_value("filetype",   "NvimTreeFilter", { buf = overlay_bufnr })
   else
     vim.api.nvim_buf_set_option(overlay_bufnr, "modifiable", true) ---@diagnostic disable-line: deprecated
+    vim.api.nvim_buf_set_option(overlay_bufnr, "filetype",   "NvimTreeFilter") ---@diagnostic disable-line: deprecated
   end
 
   vim.api.nvim_buf_set_lines(overlay_bufnr, 0, -1, false, { self.filter })
@@ -220,9 +222,9 @@ function LiveFilter:clear_filter()
   self.explorer.renderer:draw()
 
   if node then
-    utils.focus_file(node.absolute_path)
+    self.explorer:focus_node_or_parent(node)
   elseif last_node then
-    utils.focus_file(last_node.absolute_path)
+    self.explorer:focus_node_or_parent(last_node)
   end
 end