+++ lua/nvim-tree/git/init.lua
@@ -193,9 +193,10 @@ function M.get_toplevel(path)
     end
   end
 
-  -- attempt to fetch toplevel
+  -- attempt to fetch toplevel, cache if untracked
   local toplevel, git_dir = git_utils.get_toplevel(path)
   if not toplevel or not git_dir then
+    M._toplevels_by_path[path] = false
     return nil
   end
   local toplevel_norm = vim.fn.fnamemodify(toplevel, ":p")
@@ -232,7 +233,13 @@ local function reload_tree_at(toplevel)
   end
 
   log.line("watcher", "git event executing '%s'", toplevel)
-  local root_node = utils.get_node_from_path(toplevel)
+
+  local explorer = require("nvim-tree.core").get_explorer()
+  if not explorer then
+    return nil
+  end
+
+  local root_node = explorer:get_node_from_path(toplevel)
   if not root_node then
     return
   end
@@ -251,7 +258,7 @@ local function reload_tree_at(toplevel)
       end)
       :iterate()
 
-    root_node.explorer.renderer:draw()
+    explorer.renderer:draw()
   end)
 end