+++ lua/nvim-tree/view.lua
@@ -12,6 +12,9 @@ local M = {}
 
 local DEFAULT_MIN_WIDTH = 30
 local DEFAULT_MAX_WIDTH = -1
+local DEFAULT_LINES_EXCLUDED = {
+  "root",
+}
 local DEFAULT_PADDING = 1
 
 M.View = {
@@ -101,12 +104,13 @@ local function create_buffer(bufnr)
 
   local tab = vim.api.nvim_get_current_tabpage()
   BUFNR_PER_TAB[tab] = bufnr or vim.api.nvim_create_buf(false, false)
-  vim.api.nvim_buf_set_name(M.get_bufnr(), "NvimTree_" .. tab)
 
   bufnr = M.get_bufnr()
   for _, option in ipairs(BUFFER_OPTIONS) do
     vim.api.nvim_set_option_value(option.name, option.value, { buf = bufnr })
   end
+
+  vim.api.nvim_buf_set_name(M.get_bufnr(), "NvimTree_" .. tab)
 
   require("nvim-tree.keymap").on_attach(M.get_bufnr())
 
@@ -254,7 +258,6 @@ local function close(tabpage)
           return
         end
       end
-      events._dispatch_on_tree_close()
       return
     end
   end
@@ -289,6 +292,7 @@ function M.open(options)
 
   local profile = log.profile_start("view open")
 
+  events._dispatch_on_tree_pre_open()
   create_buffer()
   open_window()
   M.resize()
@@ -303,7 +307,7 @@ function M.open(options)
 end
 
 local function grow()
-  local starts_at = M.is_root_folder_visible(require("nvim-tree.core").get_cwd()) and 1 or 0
+  local starts_at = (M.is_root_folder_visible(require("nvim-tree.core").get_cwd()) and M.View.root_excluded) and 1 or 0
   local lines = vim.api.nvim_buf_get_lines(M.get_bufnr(), starts_at, -1, false)
   -- number of columns of right-padding to indicate end of path
   local padding = get_size(M.View.padding)
@@ -314,38 +318,25 @@ local function grow()
     padding = padding + wininfo[1].textoff
   end
 
-  local resizing_width = M.View.initial_width - padding
-  local max_width
-
-  -- maybe bound max
-  if M.View.max_width == -1 then
-    max_width = -1
-  else
-    max_width = get_width(M.View.max_width) - padding
+  local final_width = M.View.initial_width
+  local max_width = math.huge
+  if M.View.max_width ~= -1 then
+    max_width = get_width(M.View.max_width)
   end
 
   local ns_id = vim.api.nvim_get_namespaces()["NvimTreeExtmarks"]
   for line_nr, l in pairs(lines) do
-    local count = vim.fn.strchars(l)
+    local line_width = vim.fn.strchars(l)
     -- also add space for right-aligned icons
     local extmarks = vim.api.nvim_buf_get_extmarks(M.get_bufnr(), ns_id, { line_nr, 0 }, { line_nr, -1 }, { details = true })
-    for _, extmark in ipairs(extmarks) do
-      local virt_texts = extmark[4].virt_text
-      if virt_texts then
-        for _, virt_text in ipairs(virt_texts) do
-          count = count + vim.fn.strchars(virt_text[1])
-        end
-      end
-    end
-    if resizing_width < count then
-      resizing_width = count
-    end
-    if M.View.adaptive_size and max_width >= 0 and resizing_width >= max_width then
-      resizing_width = max_width
+    line_width = line_width + utils.extmarks_length(extmarks) + padding
+    final_width = math.max(final_width, line_width)
+    if final_width >= max_width then
+      final_width = max_width
       break
     end
   end
-  M.resize(resizing_width + padding)
+  M.resize(final_width)
 end
 
 function M.grow_from_content()
@@ -414,6 +405,7 @@ end
 ---@param opts OpenInWinOpts|nil
 function M.open_in_win(opts)
   opts = opts or { hijack_current_buf = true, resize = true }
+  events._dispatch_on_tree_pre_open()
   if opts.winid and vim.api.nvim_win_is_valid(opts.winid) then
     vim.api.nvim_set_current_win(opts.winid)
   end
@@ -425,6 +417,7 @@ function M.open_in_win(opts)
     M.reposition_window()
     M.resize()
   end
+  events._dispatch_on_tree_open()
 end
 
 function M.abandon_current_window()
@@ -605,6 +598,8 @@ function M.configure_width(width)
     M.View.adaptive_size = true
     M.View.width = width.min or DEFAULT_MIN_WIDTH
     M.View.max_width = width.max or DEFAULT_MAX_WIDTH
+    local lines_excluded = width.lines_excluded or DEFAULT_LINES_EXCLUDED
+    M.View.root_excluded = vim.tbl_contains(lines_excluded, "root")
     M.View.padding = width.padding or DEFAULT_PADDING
   elseif width == nil then
     if M.config.width ~= nil then
@@ -629,6 +624,7 @@ function M.setup(opts)
   M.View.tab = opts.tab
   M.View.preserve_window_proportions = options.preserve_window_proportions
   M.View.winopts.cursorline = options.cursorline
+  M.View.winopts.cursorlineopt = options.cursorlineopt
   M.View.winopts.number = options.number
   M.View.winopts.relativenumber = options.relativenumber
   M.View.winopts.signcolumn = options.signcolumn